+++ test/models/vrp_test.rb
@@ -147,7 +147,7 @@ module Models
           lapse: 3
         }]
 
-        generated_vrp = TestHelper.create(Oj.load(Oj.dump(vrp)))
+        generated_vrp = TestHelper.create(vrp)
         assert_equal 3, generated_vrp.relations.size
         assert_equal 2, generated_vrp.relations.first.linked_ids.size
         assert_includes generated_vrp.relations.first.linked_ids, 'shipment_0_delivery'
@@ -163,7 +163,7 @@ module Models
         linked_ids: ['service', 'shipment_1'],
         lapse: 3
       }]
-      generated_vrp = TestHelper.create(Oj.load(Oj.dump(vrp)))
+      generated_vrp = TestHelper.create(vrp)
       assert_includes generated_vrp.relations.first.linked_ids, 'service'
       assert_includes generated_vrp.relations.first.linked_ids, 'shipment_1_pickup'
 
@@ -172,7 +172,7 @@ module Models
         linked_ids: ['service', 'shipment_0', 'shipment_1'],
         lapse: 3
       }]
-      generated_vrp = TestHelper.create(Oj.load(Oj.dump(vrp)))
+      generated_vrp = TestHelper.create(vrp)
       assert_includes generated_vrp.relations.first.linked_ids, 'service'
       assert_includes generated_vrp.relations.first.linked_ids, 'shipment_0_pickup'
       assert_includes generated_vrp.relations.first.linked_ids, 'shipment_1_pickup'
@@ -181,14 +181,14 @@ module Models
         vrp[:relations].first[:type] = relation_type
         vrp[:relations].first[:linked_ids] = ['service', 'shipment_1']
         assert_raises OptimizerWrapper::DiscordantProblemError do
-          TestHelper.create(Oj.load(Oj.dump(vrp)))
+          TestHelper.create(vrp)
         end
       }
 
       vrp[:relations].first[:linked_ids] = ['service']
       %i[sequence order].each{ |relation_type|
         vrp[:relations].first[:type] = relation_type
-        TestHelper.create(Oj.load(Oj.dump(vrp))) # check no error provided
+        TestHelper.create(vrp) # check no error provided
       }
     end